refactor: move getters and setters into utils file and add OCCF
