+++ include/http-response.h
@@ -36,17 +36,20 @@ class HttpResponse
     public:
         HttpResponse();
         HttpResponse(StatusCode code);
+        HttpResponse(const HttpResponse& other);
+        HttpResponse& operator=(const HttpResponse& other);
+        ~HttpResponse();
 
         std::string reasonPhrase(StatusCode code);
         std::string buildStatusLine();
         void printResponseHeaders();
+        void appendToBody(const std::string& bodyData);
         void handleGetRequest(const std::string& uri, Server &serverConfig, const Location& location);
         void handlePostRequest(const HttpRequest& request, const std::string& locationPath);
         void clearResponse();
 
         std::string getHttpDate();
         std::string toString();
-        void appendToBody(const std::string& bodyData);
 
         void setStatus(StatusCode code);
         void setHeader(const std::string& name, const std::string& value);
