--- src/http-response/index.pdf
@@ -1,5 +1,0 @@
-<html>
-    <body>
-        <h1>Hello World</h1>
-    </body>
-</html>
